added changes that works with fine-tuned codegen models

+++ converter/huggingface_gptj_convert.py
@@ -177,7 +177,7 @@ if __name__ == "__main__":
     parser.add_argument('-trained_gpu_num', '-t_g', type=int, help='How many gpus for training', default=1)
     parser.add_argument('-infer_gpu_num', '-i_g', type=int, help='How many gpus for inference', required=True)
     parser.add_argument("-processes", "-p", type=int, help="How many processes to spawn for conversion (default: 4)", default=4)
-    parser.add_argument("-weight_data_type", type=str, default="fp16", choices=["fp32", "fp16"], help="output weight data type")
+    parser.add_argument("-weight_data_type", type=str, default="fp32", choices=["fp32", "fp16"], help="output weight data type")
 
     args = parser.parse_args()
     print("\n=============== Argument ===============")
